# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Happening API (Evently)
  version: 1.0.0
  description: |
    OpenAPI specification for the Happening/Evently backend. 
    This API demonstrates a scalable, production-grade architecture featuring:
    - **Redis Caching:** For all public GET routes to handle read-heavy traffic.
    - **RabbitMQ Queueing:** For the /bookings endpoint to handle write-heavy "flash sale" scenarios.
    - **PostgreSQL Transactions (FOR UPDATE):** To ensure data integrity and prevent overselling.
    - **Full Admin & Host Management:** Complete with approval flows and analytics.
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://your-deployed-url.com/api
    description: Production server

# ==========================================================
#  Reusable Components (Schemas & Security)
# ==========================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "User's JWT token, prefixed with 'Bearer '"
  schemas:
...
    PaymentOrderRequest:
      type: object
      properties:
        booking_id:
          type: string
          format: uuid
      required: [booking_id]
      example:
        booking_id: "b1c2d3e4-..."
    PaymentOrderResponse:
      type: object
      properties:
        order_id:
          type: string
        amount:
          type: number
        currency:
          type: string
        key_id:
          type: string
      example:
        order_id: "order_9A33XWu170gUtm"
        amount: 998.0
        currency: "INR"
        key_id: "rzp_test_XXXX"
    VerifyPaymentRequest:
      type: object
      required: [razorpay_order_id, razorpay_payment_id, razorpay_signature]
      properties:
        razorpay_order_id:
          type: string
        razorpay_payment_id:
          type: string
        razorpay_signature:
          type: string
      example:
        razorpay_order_id: "order_9A33XWu170gUtm"
        razorpay_payment_id: "pay_29QQoUBi66xm2f"
        razorpay_signature: "d6f30c..."
    HostApplicationRequest:
      type: object
      required: [phone, idType, idNumber, account_holder_name, account_number, ifsc_code, bank_name]
      properties:
        phone: { type: string }
        idType: { type: string }
        idNumber: { type: string }
        organization: { type: string }
        experience: { type: string }
        bio: { type: string }
        account_holder_name: { type: string }
        account_number: { type: string }
        ifsc_code: { type: string }
        bank_name: { type: string }
        account_type: { type: string, enum: [savings, current], default: "savings" }
...
paths:
...
  /payments/create-order:
    post:
      summary: Create a Razorpay order for a pending booking
      tags: [Payments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentOrderRequest"
      responses:
        "200":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrderResponse"
        "401":
          description: Unauthorized
        "404":
          description: Pending booking not found
  /payments/verify-payment:
    post:
      summary: Verify Razorpay payment and confirm booking
      tags: [Payments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentRequest"
      responses:
        "200":
          description: Verified
        "400":
          description: Invalid signature

  # --- Protected Host Routes ---
  /hosts/apply:
    post:
      summary: Apply to become a host (or update bank details)
      tags: [Host]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostApplicationRequest"
      responses:
        "200":
          description: Application submitted
        "400":
          description: Bad request (e.g., already pending)
        "401":
          description: Unauthorized
  /hosts/my-status:
    get:
      summary: Get the current user's host status
      tags: [Host]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized

  # --- Protected Location Routes ---
...
  /hosts/applications/pending:
    get:
      summary: (Admin) Get all pending host applications
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /hosts/applications/{userId}/review:
    post:
      summary: (Admin) Approve or reject a host application
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action: { type: string, enum: [approve, reject] }
                adminNotes: { type: string }
      responses:
        "200":
          description: OK
        "400":
          description: Invalid action
        "401":
          description: Unauthorized
        "403":
          description: Forbidden